move jsonCifToMolfile
